refactor: split the scoreboard feature
